backends/python-fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pymysql==1.1.1
cryptography==44.0.3
redis==5.2.1
python-jose[cryptography]==3.4.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic-settings==2.9.1
